nexus-platform/requirements.txt
pywebview
scapy
pandas
plotly
pygetwindow
pynput
pyautogui
pydirectinput
pywin32
paramiko
pyinstaller